recognizer/requirements.txt
opencv-python==4.5.3.56
pytesseract==0.3.8
psycopg2-binary==2.9.1
loguru==0.5.3
imutils==0.5.4
numpy==1.21.2
fire==0.4.0
kafka-python==2.0.2
pydantic==1.8.1
